# correct gemm stride for concat fusion