--- a/python/aitemplate/backend/rocm/gemm/common.py
+++ b/python/aitemplate/backend/rocm/gemm/common.py
@@ -26,12 +26,21 @@ from ...common import gemm_common
 from ...target import Target
 
 # pylint: disable=C0103,C0415,W0611,C0301
+OUTPUT_ADDR_CALCULATOR = jinja2.Template(
+    """
+  {% if output_accessor.is_from_strided_tensor %}
+    output_stride = {{output_accessor.actual_total_elements_from_stride_dim}};
+    output_offset = {{output_accessor.offset}};
+  {% endif %}
+    """
+)
 
 EXTRA_SHAPE_TEMPLATE = jinja2.Template(
     """
 {{indent}}const int64_t stride_a = *a_dim1;
 {{indent}}const int64_t stride_b = *b_dim1;
 {{indent}}const int64_t stride_c = *c_dim1;
+{{indent}}int64_t output_stride = stride_c;
 """
 )
 
@@ -42,6 +51,7 @@ INSTANCE_TEMPLATE = jinja2.Template(
 using {{name}} = {{ config_name }};
 """
 )
+
 
 EXEC_TEMPLATE = jinja2.Template(
     """
@@ -134,6 +144,7 @@ void {{function_name}}(
     hipStream_t stream
     ) {
   {{shape_func}}
+  int64_t output_offset = 0;
   {{extra_shape}}
   {{input_addr_calculator}}
   {{output_addr_calculator}}
@@ -202,7 +213,7 @@ PROBLEM_ARGS_TEMPLATE = jinja2.Template(
                                                                     static_cast<ck::half_t *>(d1_ptr)},
 {% endif %}
 {% endif %}
-{{indent}}                                static_cast<ck::half_t *>(out_ptr),
+{{indent}}                                static_cast<ck::half_t *>(out_ptr) + output_offset,
 {% if gemm_flag not in ["permute_m2n3", "bias_permute_m2n3", "bias_permute_m3n2"]  %}
 {{indent}}                                M,
 {{indent}}                                N,
@@ -235,7 +246,7 @@ PROBLEM_ARGS_TEMPLATE = jinja2.Template(
 {% elif has_d1 %}
 {{indent}}                                std::array<ck::index_t, 3>{0, static_cast<int>(stride_c), static_cast<int>(stride_c)},
 {% endif %}
-{{indent}}                                stride_c,
+{{indent}}                                output_stride,
 {% endif %}
 {{indent}}                                ck::tensor_operation::element_wise::PassThrough{},
 {{indent}}                                ck::tensor_operation::element_wise::PassThrough{},