--- a/python/aitemplate/backend/rocm/gemm/gemm_rcr.py
+++ b/python/aitemplate/backend/rocm/gemm/gemm_rcr.py
@@ -90,7 +90,9 @@ def gemm_gen_function(func_attrs, exec_cond_template, dim_info_dict):
     str
         The rendered template of generated function body.
     """
-    return common.gen_function(func_attrs, exec_cond_template, dim_info_dict, "")
+    return common.gen_function(func_attrs, exec_cond_template, dim_info_dict, "",
+        input_addr_calculator=common.INPUT_ADDR_CALCULATOR.render(accessor_a=func_attrs["input_accessors"][0], accessor_b=func_attrs["input_accessors"][1]),
+        output_addr_calculator=common.OUTPUT_ADDR_CALCULATOR.render(output_accessor=func_attrs["output_accessors"][0]))
 
 
 @registry.reg("rocm.gemm_rcr.func_decl")