--- a/python/aitemplate/backend/rocm/gemm/gemm_rcr_bias.py
+++ b/python/aitemplate/backend/rocm/gemm/gemm_rcr_bias.py
@@ -90,7 +90,7 @@ def gemm_gen_function(func_attrs, exec_cond_template, dim_info_dict):
     str
         The rendered template of generated function body.
     """
-    return common.gen_function(func_attrs, exec_cond_template, dim_info_dict, "bias")
+    return common.gen_function(func_attrs, exec_cond_template, dim_info_dict, "bias", output_addr_calculator=common.OUTPUT_ADDR_CALCULATOR.render(output_accessor=func_attrs["output_accessors"][0]))
 
 
 @registry.reg("rocm.gemm_rcr_bias.func_decl")