--- a/python/aitemplate/backend/rocm/gemm/gemm_rcr_bias_sigmoid_mul.py
+++ b/python/aitemplate/backend/rocm/gemm/gemm_rcr_bias_sigmoid_mul.py
@@ -129,6 +129,8 @@ def gen_function(
         dim_info_dict,
         "bias_sigmoid_mul",
         extra_code=EXTRA_CODE.render(),
+        input_addr_calculator=common.INPUT_ADDR_CALCULATOR.render(accessor_a=func_attrs["input_accessors"][0], accessor_b=func_attrs["input_accessors"][1]),
+        output_addr_calculator=common.OUTPUT_ADDR_CALCULATOR.render(output_accessor=func_attrs["output_accessors"][0]),
     )