--- a/python/aitemplate/backend/rocm/gemm/gemm_rcr_bias_swish.py
+++ b/python/aitemplate/backend/rocm/gemm/gemm_rcr_bias_swish.py
@@ -96,6 +96,8 @@ def gemm_gen_function(func_attrs, exec_cond_template, dim_info_dict):
         exec_cond_template,
         dim_info_dict,
         "bias_swish",
+        input_addr_calculator=common.INPUT_ADDR_CALCULATOR.render(accessor_a=func_attrs["input_accessors"][0], accessor_b=func_attrs["input_accessors"][1]),
+        output_addr_calculator=common.OUTPUT_ADDR_CALCULATOR.render(output_accessor=func_attrs["output_accessors"][0]),
     )