--- a/python/aitemplate/compiler/transform/transform_strided_ops.py
+++ b/python/aitemplate/compiler/transform/transform_strided_ops.py
@@ -479,6 +479,7 @@ def transform_strided_ops(
         funcs = [
             # Keep on ROCM
             _fuse_strided_op_and_view_op,
+            _fuse_strided_op_and_cat,
             _fuse_split_and_strided_op,
             _fuse_slices_concat,
         ]