--- a/python/aitemplate/compiler/transform/transform_strided_ops.py
+++ b/python/aitemplate/compiler/transform/transform_strided_ops.py
@@ -481,6 +481,7 @@ def transform_strided_ops(
             _fuse_strided_op_and_view_op,
             _fuse_strided_op_and_cat,
             _fuse_split_and_strided_op,
+            _fuse_slice_and_strided_op,
             _fuse_slices_concat,
         ]
     for func in funcs: