Bert performance(old):
output_0 shape: [32, 1024, 768]

batch_size: 32, seq_length: 1024, latency: 89.88082313537598

output_0 shape: [64, 1024, 768]

batch_size: 64, seq_length: 1024, latency: 181.42297744750977

output_0 shape: [128, 1024, 768]

batch_size: 128, seq_length: 1024, latency: 357.5509796142578

output_0 shape: [256, 1024, 768]

batch_size: 256, seq_length: 1024, latency: 745.2749176025391

Bert performance(new):
output_0 shape: [32, 1024, 768]

batch_size: 32, seq_length: 1024, latency: 89.81751441955566

output_0 shape: [64, 1024, 768]

batch_size: 64, seq_length: 1024, latency: 180.47053909301758

output_0 shape: [128, 1024, 768]

batch_size: 128, seq_length: 1024, latency: 357.5366973876953

output_0 shape: [256, 1024, 768]

batch_size: 256, seq_length: 1024, latency: 743.6724853515625


Unit performance tests
slice op fusion 0.01985735073685646ms->0.015206613577902317 ms
layernorm fusion 0.0848146378993988ms->0.08335678279399872ms