@carlushuang 2022-11-11
differ till https://github.com/ROCmSoftwarePlatform/AITemplate/pull/15